feat(npc): migrate Jiddos NPCs to the Evil Heroes system and add Crystal Server compatibility
This update migrates legacy NPCs from the old Jiddos system to the new Evil Heroes NPC system and adds compatibility so Crystal Server NPC scripts can now work on our TFS.

It becomes much easier to port full NPCs from Canary, Crystal Server, or similar bases without rewriting everything from scratch. Some NPCs may still require extra storages, helper functions, or small support adjustments depending on their logic.

If an imported NPC uses storages for quests, travel, bank logic, missions, promotions, blessings, or other state-based interactions, you may need to add the required entries in data/lib/storages.lua and include any extra NPC-specific support code needed for it to work correctly in our TFS environment.

+++ src/protocolgame.cpp
@@ -3232,8 +3232,8 @@ void ProtocolGame::AddShopItem(NetworkMessage& msg, const ShopInfo& item)
 
 	msg.addString(item.realName);
 	msg.add<uint32_t>(it.weight);
-	msg.add<uint32_t>(std::max<uint32_t>(item.buyPrice, 0));
-	msg.add<uint32_t>(std::max<uint32_t>(item.sellPrice, 0));
+	msg.add<uint32_t>(static_cast<uint32_t>(std::max<int64_t>(item.buyPrice, 0)));
+	msg.add<uint32_t>(static_cast<uint32_t>(std::max<int64_t>(item.sellPrice, 0)));
 }
 
 void ProtocolGame::parseExtendedOpcode(NetworkMessage& msg)
